fix(pipelineTriggers): handle invalid constraint regexes (#1209)
* fix(pipelineTriggers): handle invalid constraint regexes

Currently, if a user accidentally configures a pipeline trigger with an
invalid regex for a constraint value, echo throws a
`PatternSyntaxException` any time it receives an event that should be
tested against the constraint.  As a result, all pipeline triggers using
the same trigger type & source/subscription will be non-functional, not
just the pipeline with the bad regex.

This patch catches `PatternSyntaxException`, logs an error about the bad
regex for troubleshooting purposes, and returns `false` such that the
bad regex will never match, allowing all other pipeline triggers to
continue functioning normally.

* fix(pipelineTriggers): improved logging for invalid constraint patterns